Modularized into logical components
